requirements.txt
flask_sqlalchemy
flask_migrate
flask_wtf
flask_uploads
flask_login
email_validator
flask_script